% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{course_data}
\alias{course_data}
\title{Metadata for all Smith College Spring 2024 courses (thus far)}
\format{
A data frame with 763 rows and 15 columns
\describe{
\item{course_num}{Course number}
\item{course_dept}{Department the course is in}
\item{course_sub}{Course subject}
\item{course_section}{Section of the course}
\item{course_name}{Name of the course}
\item{course_instructor}{Instructor of the course}
\item{course_status}{Status of the course (i.e. if it is open for registration)}
\item{credits}{Credits of the course}
\item{transcript}{Transcription of the text in the comic}
\item{max_enroll}{Maximun number of enrollment}
\item{waitlist}{Current number of people on the waitlist}
\item{reserved}{If there are reserved seats in this course}
\item{meeting_time}{Time of course sessions}
\item{description}{Course description}
\item{course_id}{A unique identifier that is created by pasting together course_sub, course_num and course_section}
}
}
\source{
\url{https://www.smith.edu/apps/course_search/}
}
\usage{
course_data
}
\description{
Metadata for all Smith College Spring 2024 courses (thus far)
}
\keyword{datasets}
