% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Course_rec.R
\name{course_description}
\alias{course_description}
\title{Description of desired course.}
\usage{
course_description(course, data)
}
\arguments{
\item{course}{The specific course id.}

\item{data}{The dataset of course offerings to be searched}
}
\value{
The description of the course that refers to the course id input.
\itemize{
\item description
}
}
\description{
Given a specific course id, this function returns a description of that course.
}
\examples{
# Returns course description for the AFR11701 class.
course_description("AFR11701", course_data)

}
