% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Course_rec.R
\name{course_meeting_time}
\alias{course_meeting_time}
\title{Meeting times for courses offered}
\usage{
course_meeting_time(course, data)
}
\arguments{
\item{course_id}{The specific course id for the course the user wants to know the meeting time of.}
}
\value{
The meeting time that corresponds to the given course id.
}
\description{
Given a course_id, the function returns the meeting time for that specific course.
}
