% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Course_rec.R
\name{course_recommend}
\alias{course_recommend}
\title{Generate all available courses}
\usage{
course_recommend(course1, course2, course3, data = course_data_na_removed)
}
\arguments{
\item{course1}{the first course the user plans to take}

\item{course2}{the second course the user plans to take}

\item{course3}{the third course the user plans to take}

\item{data}{Data set containing entire course schedule. The default data set contains course schedule of
all spring 2024 courses at Smith College.}
}
\value{
A character vector of the course IDs of all available courses, meaning courses that do not have a
time conflict with what the user's three courses.
}
\description{
Given three courses the user plans to take,
this functions returns a character vector of all courses available to the user.
}
