% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Course_rec.R
\name{course_recommend_dept}
\alias{course_recommend_dept}
\title{A function which allows the user to specify a department they would like to take
a class in, e.g. MTH for the Math department}
\usage{
course_recommend_dept(courses, dept, data)
}
\arguments{
\item{courses}{a vector of three course ids}

\item{dept}{the three letter department code which the user wants to search}

\item{data}{the data set of course offerings to be searched}
}
\value{
a data frame of all possible classes within the specified department which have no overlap with the entered classes
}
\description{
A function which allows the user to specify a department they would like to take
a class in, e.g. MTH for the Math department
}
\examples{
course_rec_dept(c('AFR11701', 'AFR17501', 'AFR24901'), 'MTH', course_data)

}
