% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Course_rec.R
\name{course_recommend_exlude_day}
\alias{course_recommend_exlude_day}
\title{Function which allows the user to optionally enter a day that will be excluded from the search (any classes that meet that day}
\usage{
course_recommend_exlude_day(courses, exclude_day, dept, data)
}
\arguments{
\item{courses}{a character vector of three course ids}

\item{exclude_day}{the day of the week the user wants to exclude from the search. any classes which meet on this day will NOT be returned}

\item{dept}{the three letter department code which the user wants to search}

\item{data}{the dataset of course offerings to be searched}
}
\value{
a dataframe containing all the classes the user can take with the corresponding optional criteria
}
\description{
Function which allows the user to optionally enter a day that will be excluded from the search (any classes that meet that day
}
\examples{
course_rec_exclude_day_dept(c('AFR11701', 'AFR17501', 'AFR24901'), 'Monday', 'MTH', course_data)

}
