% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Course_rec.R
\name{course_schedule}
\alias{course_schedule}
\title{Generate a course schedule for 3 courses}
\usage{
course_schedule(course1, course2, course3, data = course_data_na_removed)
}
\arguments{
\item{course1}{the first course}

\item{course2}{the second course}

\item{course3}{the third course}

\item{data}{A data set containing entire course schedule. The default data set contains course schedule of
all spring 2024 courses at Smith College.}
}
\value{
A data frame containing one column, which is a character vector of three meeting times
}
\description{
Given three character vectors of course IDs, this function returns their meeting times as a column in a data frame
}
