% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Course_rec.R
\name{find_overlap}
\alias{find_overlap}
\title{Find if there's a time conflict between two fine grained schedules}
\usage{
find_overlap(a, b)
}
\arguments{
\item{a}{the first fine grained schedule}

\item{b}{the second fine grained schedule}
}
\value{
A logical vector. TRUE if there is a time conflict, FALSE if not.
}
\description{
Given two fine grained schedules, this function checks if they have any time conflicts
}
