% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Course_rec.R
\name{fine_grained_schedule}
\alias{fine_grained_schedule}
\title{Generate a fine grained schedule for course meeting times.}
\usage{
fine_grained_schedule(x)
}
\arguments{
\item{x}{A character vector of course meeting times.}
}
\value{
A list of fine grained schedules indicating one start time and one end time for each weekday.
}
\description{
Given a character vector of course meeting times, this function returns a fine grained schedule.
}
