% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Course_rec.R
\name{remove_locations}
\alias{remove_locations}
\title{Clean meeting_time column}
\usage{
remove_locations(x)
}
\arguments{
\item{x}{the meeting time column from course data set}
}
\value{
cleaned meeting time column only containing meeting time
}
\description{
Removes locations from the meeting time column of course data data set
}
