% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Course_rec.R
\name{scrap_course_data}
\alias{scrap_course_data}
\title{Scrap course data from Smith College course search website}
\usage{
scrap_course_data(year, semester)
}
\arguments{
\item{year}{A scalar numeric vector indicating the year}

\item{semester}{A character vector indicating whether the semester is the fall, spring or interterm semester}
}
\value{
data set
}
\description{
Given the year and semester, creates a data set containing info from all courses in that semester from website
}
