% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Course_rec.R
\name{subject_rec}
\alias{subject_rec}
\title{Gives reccomendation for user who wants to take a course with a specific course subject code.}
\usage{
subject_rec(subject_code)
}
\arguments{
\item{subject_code}{the course subject the user wants to find classes offered in.}
}
\value{
All of the courses offered with the given course subject code.
}
\description{
Given a subject code, this returns all courses with that subject code offered at Smith college.
}
