% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Course_rec.R
\name{course_description}
\alias{course_description}
\title{Description of desired course.}
\usage{
course_description(course, data)
}
\arguments{
\item{course_id}{The specific course id}
}
\value{
The description of the course that refers to the course id input.
\itemize{
\item description
}
}
\description{
Given a specific course id, this function returns a description of that course.
}
