% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Course_rec.R
\name{course_recommend}
\alias{course_recommend}
\title{Generate all available courses}
\usage{
course_recommend(courses, data)
}
\arguments{
\item{courses}{a character vector of length 3 containing the course IDs of all 3 courses the user plans to take.}

\item{data}{Data set containing entire course schedule.}
}
\value{
A character vector of the course IDs of all available courses, meaning courses that do not have a
time conflict with what the user's three courses.
}
\description{
Given three courses the user plans to take,
this functions returns a character vector of all courses available to the user.
}
\examples{
course_recommend(c('AFR11701', 'AFR17501', 'AFR24901'), course_data)

}
