% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Course_rec.R
\name{course_time}
\alias{course_time}
\title{Generate the meeting time of a course}
\usage{
course_time(course, data = course_data_na_removed)
}
\arguments{
\item{course}{A character vector identifying a course ID.}

\item{data}{A data set containing the entire course schedule. The default data set contains the course schedule of
all spring 2024 courses at Smith College.}
}
\value{
A character vector of the meeting time of the course.
}
\description{
Given a course ID and a data set of all courses, this function returns the meeting time of the inputted course as a character string,
if the course ID exists in courses data set.
}
